Boot3 (#78)
* konquest_time.c

* portability changes

* prototype fix for mwfile_init_for_mk

* improments in konquest_nav.c

* portability and type corrections in function bodies and signatures

* more cleanups and cast removal

+++ include/mw/mwFile.h
@@ -9,16 +9,16 @@ typedef enum mwTargetMemAlign {
 } mwTargetMemAlign;
 
 typedef struct mwFileInitParam {
-    unsigned long field_0x00;
+    unsigned int field_0x00;
     short max_open_files;       /* +0x04 */
     short field_0x06;
     short field_0x08;
     short field_0x0A;
-    unsigned long flags;        /* +0x0C */
+    unsigned int flags;         /* +0x0C */
     short field_0x10;
     short field_0x12;
     void* allocator_context;    /* +0x14 */
-    unsigned long field_0x18;
+    unsigned int field_0x18;
 } mwFileInitParam; /* 0x1C */
 
 typedef int (*mwFileErrorCallback)(int operation, int error);
@@ -26,7 +26,7 @@ typedef int (*mwFileErrorCallback)(int operation, int error);
 typedef union mwFileAsyncValue {
     void* pointer;
     _mwFile* file;
-    unsigned long bytes;
+    unsigned int bytes;
 } mwFileAsyncValue;
 
 typedef struct _mwFileAsyncResult {
@@ -56,10 +56,10 @@ mwFileCommand* mwFileOpenAsync(const char* path, int flags,
 mwFileCommand* mwFileCloseAsync(_mwFile* file, int flags,
                                 mwFileCallback callback);
 mwFileCommand* mwFileReadAsync(_mwFile* file, long long offset, void* buffer,
-                               unsigned long length, int count,
+                               unsigned int length, int count,
                                mwFileCallback callback, void* arg);
 mwFileCommand* mwFileWriteAsync(_mwFile* file, long long offset, void* buffer,
-                                unsigned long length, int count,
+                                unsigned int length, int count,
                                 mwFileCallback callback, void* arg);
 unsigned char mwFileIsCommandCompleted(mwFileCommand* command,
                                        mwFileAsyncResult* result);
